Frist version of letter + some style updates
This adds a page for the „Thoughts on Office-Bound Work“ letter under /hotnews/thoughts-on-office-bound-work. As the letter is very long, it’s easier to manage in Markdown, so I added a Markdown plugin, so we can write in Markdown directly instead of having to deal with formatting everything with HTML.

I also extracted the footer that was reused on all the pages into a component and added the new letter to the list of links there.